third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/block-local-documents-inheriting-none.https-expected.txt

This is a testharness.js-based test.
FAIL Prevent local scheme documents from loading within a COEP: require-corp iframe if they inherit COEP: none assert_equals: expected 3 but got 0
Harness: the test ran to completion.
